Change my Gradle file

+++ src/main/antlr/ExprPostOrderVisitor.java
@@ -6,7 +6,6 @@ import logic.*;
 
 import java.util.ArrayDeque;
 import java.util.Queue;
-import java.math.BigDecimal;
 import org.antlr.v4.runtime.tree.AbstractParseTreeVisitor;
 
 /**
@@ -39,9 +38,9 @@ public class ExprPostOrderVisitor extends AbstractParseTreeVisitor<Queue<Atom>> 
 		visitChildren(ctx);
 		if (ctx.getChildCount() > 1 ) {
 			if (ctx.getChild(1).getText().equals("+")) {
-				operations.add(new Atom(false, 0, Atom.Function.PLUS));
+				operations.add(new Atom(false, (double)0, Atom.Function.PLUS));
 			} else if (ctx.getChild(1).getText().equals("-")) {
-				operations.add(new Atom(false, 0, Atom.Function.MINUS));
+				operations.add(new Atom(false, (double)0, Atom.Function.MINUS));
 			}
 		}
 		return null;
@@ -56,9 +55,9 @@ public class ExprPostOrderVisitor extends AbstractParseTreeVisitor<Queue<Atom>> 
 		visitChildren(ctx);
 		if (ctx.getChildCount() > 1) {
 			if (ctx.getChild(1).getText().equals("*")) {
-				operations.add(new Atom(false, 0, Atom.Function.MULTIPLY));
+				operations.add(new Atom(false, (double)0, Atom.Function.MULTIPLY));
 			} else if (ctx.getChild(1).getText().equals("/")) {
-				operations.add(new Atom(false, 0, Atom.Function.DIVIDE));
+				operations.add(new Atom(false, (double)0, Atom.Function.DIVIDE));
 			}
 		}
 		return null;
@@ -72,7 +71,7 @@ public class ExprPostOrderVisitor extends AbstractParseTreeVisitor<Queue<Atom>> 
 	@Override public Queue<Atom> visitExpr_pw(ExprParser.Expr_pwContext ctx) {
 		visitChildren(ctx);
 		if (ctx.getChildCount() > 1) {
-				operations.add(new Atom(false, 0, Atom.Function.POWER));
+				operations.add(new Atom(false, (double) 0, Atom.Function.POWER));
 		}
 		return null;
 	}
@@ -84,7 +83,7 @@ public class ExprPostOrderVisitor extends AbstractParseTreeVisitor<Queue<Atom>> 
 	 */
 	@Override public Queue<Atom> visitExpr_func_mul(ExprParser.Expr_func_mulContext ctx) {
 		visitChildren(ctx);
-		operations.add(new Atom(false, 0, Atom.Function.MULTIPLY));
+		operations.add(new Atom(false, (double)0, Atom.Function.MULTIPLY));
 		return null;
 	}
 	@Override public Queue<Atom> visitExpr_func(ExprParser.Expr_funcContext ctx) {
@@ -94,32 +93,32 @@ public class ExprPostOrderVisitor extends AbstractParseTreeVisitor<Queue<Atom>> 
 			String factorial_func = ctx.getChild(1).getText();
 			switch (func) {
 				case "sin(":
-					operations.add(new Atom(false, 0, Atom.Function.SIN));
+					operations.add(new Atom(false, (double)0, Atom.Function.SIN));
 					break;
 				case "cos(":
-					operations.add(new Atom(false, 0, Atom.Function.COS));
+					operations.add(new Atom(false, (double)0, Atom.Function.COS));
 					break;
 				case "tan(":
-					operations.add(new Atom(false, 0, Atom.Function.TAN));
+					operations.add(new Atom(false, (double)0, Atom.Function.TAN));
 					break;
 				case "log(":
-					operations.add(new Atom(false, 0, Atom.Function.LOG));
+					operations.add(new Atom(false, (double)0, Atom.Function.LOG));
 					break;
 				case "ln(":
-					operations.add(new Atom(false, 0, Atom.Function.LN));
+					operations.add(new Atom(false, (double)0, Atom.Function.LN));
 					break;
 				case "tanh(":
-					operations.add(new Atom(false, 0, Atom.Function.TANH));
+					operations.add(new Atom(false, (double)0, Atom.Function.TANH));
 					break;
 				case "cosh(":
-					operations.add(new Atom(false, 0, Atom.Function.COSH));
+					operations.add(new Atom(false, (double)0, Atom.Function.COSH));
 					break;
 				case "sinh(":
-					operations.add(new Atom(false, 0, Atom.Function.SINH));
+					operations.add(new Atom(false, (double)0, Atom.Function.SINH));
 					break;
 			}
 			if (factorial_func.equals("!")) {
-				operations.add(new Atom(false, 0, Atom.Function.FACTORIAL));
+				operations.add(new Atom(false, (double)0, Atom.Function.FACTORIAL));
 			}
 		}
 		return null;
@@ -134,10 +133,10 @@ public class ExprPostOrderVisitor extends AbstractParseTreeVisitor<Queue<Atom>> 
 		visitChildren(ctx);
 		if (ctx.getChildCount() == 2) {
 			if (ctx.getChild((0)).getText().equals("-")){
-				operations.add(new Atom(false, 0, Atom.Function.NEG));
+				operations.add(new Atom(false, (double)0, Atom.Function.NEG));
 			}
 			if (ctx.getChild((0)).getText().equals("+")){
-				operations.add(new Atom(false, 0, Atom.Function.POS));
+				operations.add(new Atom(false, (double)0, Atom.Function.POS));
 			}
 		}
 		return null;
